Commited before moving to new repo
Commited before moving to new repo

+++ src/main/java/com/garage/garageManager/repository/dao/JobCardEntityDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JobCardEntityDao extends JpaRepository<JobCardEntity,Integer> {
+public interface JobCardEntityDao extends JpaRepository<JobCardEntity,Long> {
 }
