+++ src/main/java/com/garage/garageManager/repository/dao/ServiceDetailsDao.java
@@ -4,5 +4,5 @@ package com.garage.garageManager.repository.dao;
 import com.garage.garageManager.repository.entity.ServiceDetailsEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceDetailsDao extends JpaRepository<ServiceDetailsEntity,Integer> {
+public interface ServiceDetailsDao extends JpaRepository<ServiceDetailsEntity,Long> {
     }
