+++ src/main/java/com/garage/garageManager/repository/dao/VehicleDetailsDao.java
@@ -3,5 +3,5 @@ package com.garage.garageManager.repository.dao;
 import com.garage.garageManager.repository.entity.VehicleDetailsEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehicleDetailsDao extends JpaRepository<VehicleDetailsEntity,Integer> {
+public interface VehicleDetailsDao extends JpaRepository<VehicleDetailsEntity,Long> {
 }
